requirements.txt
aiohttp==3.8.1
Flask==1.1.4
gspread==5.2.0
gunicorn==20.1.0
google-cloud-secret-manager==2.9.2
oauth2client==4.1.3
markupsafe==2.0.1